Fixed app crashing on selecting primary Camera
When clicking the checkbox to set the primary camera outside the default scene, the app would crash.
This is now fixed, but there is another bug to fix.

If in the default scene you select the default camera and set it as primary camera, if you don't unselect it and open another scene, when you select any object the app crashes.